style: rename constants folder to enums gf-13
